Modify logging code
