+++ dxc/ai/__init__.py
@@ -46,9 +46,9 @@ from .deep_learning.image_classifier import seggregate_data
 from .deep_learning.image_classifier import split_normalize_data
 from .deep_learning.image_classifier import image_classifier
 
-from .logging.logging import pipeline_log
-from .logging.logging import experiment_design_log
-from .logging.logging import microservice_design_log
+from .logging.pipeline_logging          import pipeline_log
+from .logging.experiment_design_logging import experiment_design_log
+from .logging.microservice_logging      import microservice_design_log
 
 from .datasets._base import load_data
 from .datasets._base import load_data_details
